requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
mindsdb-sdk>=3.4.8
pydantic>=1.10.0
python-dotenv